Follow up fixes

+++ Code/UnitTests/CoreTest/Utils/IntervalSchedulerTest.cpp
@@ -211,7 +211,7 @@ EZ_CREATE_SIMPLE_TEST(Utils, IntervalScheduler)
     for (ezUInt32 i = 0; i < 32; ++i)
     {
       auto& work = works.ExpandAndGetRef();
-      work.m_IntervalMs = (i & 1);
+      work.m_IntervalMs = static_cast<float>((i & 1u));
 
       scheduler.AddOrUpdateWork(&work, ezTime::Milliseconds(i));
     }
